backend/requirements.txt
amqp==5.1.0
asgiref 
Babel==2.11.0
billiard==3.6.4.0
celery==5.2.3
certifi 
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django 
django-allauth==0.50.0
django-braces==1.15.0
django-filter==21.1
django-oauth-toolkit==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2

docopt==0.6.2
drf-social-oauth2==1.2.1
ecdsa==0.18.0
idna==3.3
install==1.3.5
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.4.2
kombu==5.2.4
Markdown==3.3.6
MarkupSafe==2.1.1
nexmo==2.5.2
oauth2-provider==0.0
oauthlib==3.2.0
phonenumberslite==8.13.4
Pillow==9.1.0
pipreqs==0.4.11
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyopt==0.84
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python3-openid==3.2.0
pytz
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse 
twilio==7.7.1
typing_extensions 
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.14.1
yarg==0.1.9
